+++ R/seed.R
@@ -10,7 +10,7 @@
 #' having applied it.
 #' @return The seed value is returned invisibly
 #' @export
-random_seed <- function(set.seed = TRUE) {
+random_seed <- function(set.seed = TRUE) { # TODO deprecate in favor of local_seed
     digits <- 9
     newSeed <- as.numeric(Sys.time()) * 10^(digits - 3)
     newSeed <- as.integer(round(newSeed, -digits) - newSeed)
@@ -18,4 +18,76 @@ random_seed <- function(set.seed = TRUE) {
         set.seed(newSeed)
     }
     return(invisible(newSeed))
+}
+
+
+
+
+
+#' @name local_seed
+#' @title Set a seed local to a call
+#' @description
+#' When used inside a call, `local_seed()` first records the existing seed
+#' then transiently sets a specific one. When the call exits, the recorded seed
+#' is set again, so that the transient seed setting leaves no effects.
+#' @param seed seed value to set
+#' @examples
+#' f <- function() {
+#'   local_seed(1234)
+#'   r_val <- rnorm(1)
+#'   return(r_val)
+#' }
+#'
+#' rnorm(1) # make sure a seed exists (not needed but handy for this example)
+#'
+#' seed1 <- .Random.seed
+#' x <- f()
+#' seed2 <- .Random.seed
+#' y <- rnorm(1)
+#' seed3 <- .Random.seed
+#' z <- f()
+#'
+#' identical(seed1, seed2)
+#' !identical(seed1, seed3)
+#'
+#' identical(x, z)
+#' !identical(x, y)
+#' @export
+local_seed <- function(seed) {
+  prev_seed <- if (.has_seed()) {
+    get(".Random.seed", 1)
+  } else {
+    NULL
+  }
+
+  .gutils_prev_seed <- NULL
+  assign(".gutils_prev_seed", prev_seed, sys.frame(-1)) # send to prev stack frame
+
+  set.seed(seed)
+
+  do.call(
+    "on.exit",
+    args = list(add = TRUE, expr = {
+      quote(
+        if (is.null(.gutils_prev_seed)) {
+          .rm_seed()
+        } else {
+          assign(".Random.seed", .gutils_prev_seed, 1)
+        }
+      )
+    }),
+    envir = sys.frame(-1)
+  )
+}
+
+
+# internals ####
+
+# based on the has_seed() and rm_seed() internals from package withr
+.has_seed <- function() {
+  exists(".Random.seed", globalenv(), mode = "integer", inherits = FALSE)
+}
+.rm_seed <- function() {
+  if (!.has_seed()) return(NULL)
+  rm(".Random.seed", envir = globalenv())
 }